requirements.txt
Django==3.2.4
django-cors-headers==4.3.1
python-decouple==3.8
django-pgtrigger==4.11.1
django-user-agents==0.4.0
psycopg2-binary==2.9.9
psycopg2==2.9.9
django-phonenumbers~=1.0.1
django-phonenumber-field~=7.3.0
djangorestframework~=3.12.4
django-allauth~=0.63.3
django-allauth[socialaccount]~=0.63.3
PyJWT~=2.8.0
cryptography~=42.0.8
python-dateutil~=2.9.0